Refactor oauth module
